chore: Member 관련 클래스를 Auth 관련 클래스로 rename
Member와 Auth의 기능을 분리하기 위해 Member 관련 클래스를
Auth 관련 클래스로 일괄 변경하였습니다
