+++ src/main/resources/mybatis/mapper/AuthMapper.xml
@@ -3,14 +3,14 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.souljit2.selleverything.member.MemberMapper">
-    <insert id="signUp" parameterType="com.souljit2.selleverything.member.MemberDTO">
+<mapper namespace="com.souljit2.selleverything.auth.AuthMapper">
+    <insert id="signUp" parameterType="com.souljit2.selleverything.model.MemberDTO">
         INSERT INTO tb_member
             (memberId, memberPassword, memberName, memberNickname, memberPhone, membershipAgreementYn, optionalInfoAgreementYn)
         VALUES (#{memberId}, #{memberPassword}, #{memberName}, #{memberNickname}, #{memberPhone}, #{membershipAgreementYn}, #{optionalInfoAgreementYn})
     </insert>
 
-    <select id="signIn" resultType="com.souljit2.selleverything.member.MemberDTO">
+    <select id="signIn" resultType="com.souljit2.selleverything.model.MemberDTO">
         SELECT id, memberPassword
         FROM tb_member
         WHERE memberId = #{memberId}
